backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
channels==4.0.0
channels-redis==4.1.0
celery==5.3.4
redis==5.0.1
daphne==4.0.0
python-decouple==3.8
mysqlclient==2.2.0
dj-database-url==2.1.0
drf-spectacular==0.27.0
django-ratelimit==4.1.0 